Fix copy-pasted variables
without var-sync we don't get the name, but at least it doesn't break.

also added error boundaries around inputs. i thiiink i broke one of my noggins with an intermediate version of this code -- hopefully not a persistent issue -- but at least with the error boundary it's possible to update the budget if your noggin breaks