Add flux models with boolean variables
lots of work to get boolean params working. still need to get the checkboxes in 'model parameters' working, but i think most everything else is done.

lord i need TESTS on this stuff ...... sdlkfjsdlkfj
like, had some weird stuff going on with boolean variables behaving differently in the I/O viz and the actual request sent to the model, which is soo annoying bc the point of the I/O viz is... visibility, lol. fixed that bug of course but no tests to check these things on a broader scale. variables code is pretty messy, i *think* out of necessity but it still makes me uncomfortable